server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
httpx>=0.26.0
aiogram>=3.15.0
PyJWT>=2.9.0
aiohttp>=3.11.0
